Avoid potential conflict
