+++ src/rockets/rockets.js
@@ -1,31 +1,18 @@
-const FETCH_ROCKETS = 'rockets/getRockets';
-const FETCH_ROCKETS_FULFILLED = 'rockets/getRockets/fulfilled';
+/* eslint-disable react-hooks/rules-of-hooks */
+/* eslint-disable no-unused-vars */
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 
-export function fetchRockets(payload) {
-  return {
-    type: FETCH_ROCKETS,
-    payload,
-  };
-}
+const Rocket = () => {
+  const rocketList = useSelector((state) => state.rockets);
+  return rocketList.map((rocket) => (
+    <div key={rocket.id}>
+      <div>{rocket.id}</div>
+      <div>{rocket.name}</div>
+      <div>{rocket.type}</div>
+      <div>{rocket.images}</div>
+    </div>
+  ));
+};
 
-export default function reducer(state = [], action = {}) {
-  let result = [];
-  switch (action.type) {
-    case FETCH_ROCKETS_FULFILLED:
-      action.payload.forEach((element) => {
-        result = [
-          ...result,
-          {
-            id: element.id,
-            name: element.rocket_name,
-            type: element.rocket_type,
-            images: element.flickr_images,
-          }];
-      });
-      return [...result];
-    default:
-      return [...state];
-  }
-}
-
-export { FETCH_ROCKETS, FETCH_ROCKETS_FULFILLED };
+export default Rocket;
